gutted binders, now adding molecule module to be the tool for small molecules, there is a design conflict here between the structure module and the molecule. I'm thinking about using the structure module to have ligand bound states in the molecule class instances
