+++ requirements.txt
@@ -2,7 +2,7 @@ pandas==2.1.4
 biopython
 torch
 beautifulsoup4
-transformers==4.51.3
+transformers==4.52.1
 qwen_vl_utils
 accelerate
 datasets
@@ -24,7 +24,7 @@ lxml
 pgvector
 adapters
 colpali-engine>=0.3.0
-omnipath[graph]
 pytest
 docker
-jsonschema
+openpyxl
+model2vec
